Fix user del

+++ src/Magend/UserBundle/Controller/ConsumerController.php
@@ -2,6 +2,7 @@
 
 namespace Magend\UserBundle\Controller;
 
+use Exception;
 use Magend\BaseBundle\Controller\BaseController as Controller;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\RedirectResponse;
@@ -47,10 +48,13 @@ class ConsumerController extends Controller
         $em = $this->getDoctrine()->getEntityManager();
         $repo = $this->getDoctrine()->getRepository('MagendUserBundle:User');
         $user = $repo->find($id);
+        if ($user->hasRole('ROLE_ADMIN')) {
+            throw new Exception('admin.no_del');
+        }
         
         $em->remove($user);
         $em->flush();
-        return $this->redirect($this->generateUrl('user_list'));
+        return $this->redirect($this->generateUrl('consumer_list'));
     }
     
     /**
